Update README.md
The program in this repository is an implementation of algorithms developed in the following paper:
Rastpour, A. A. Ingolfsson, B. Sandıkcı. (2021) Algorithms for queueing systems with reneging and priorities modeled as quasi-birth-death processes. INFORMS Journal on Computing; Under Minor Revision.

This program calculates bounds on the Class-2 Wait Probability of an Erlang A system with two priority classes. This program includes 12 MATLAB m files. A user needs to input system parameters in the main.m file and adjust the function_class2waiting.m file in case a performance measure other than the Class-2 Wait Probability is needed.